Refactored old files and touched up code comments
