+++ src/main/java/QUT/CAB302/fortunecookie/ApplicationMain.java
@@ -6,6 +6,7 @@ import javafx.scene.Scene;
 import javafx.stage.Stage;
 import java.io.IOException;
 
+// Main class and entrypoint for the application
 public class
 ApplicationMain extends Application {
     // Constants defining the window title and size
@@ -13,6 +14,7 @@ ApplicationMain extends Application {
     public static final int WIDTH = 640;
     public static final int HEIGHT = 360;
 
+    // Loads the Login layout file, sets the scene height, width, and title
     @Override
     public void start(Stage stage) throws IOException {
         FXMLLoader fxmlLoader = new FXMLLoader(ApplicationMain.class.getResource("login.fxml"));
@@ -22,6 +24,7 @@ ApplicationMain extends Application {
         stage.show();
     }
 
+    // Main method, launches the application
     public static void main(String[] args) {
 
         launch();
