+++ src/main/java/QUT/CAB302/fortunecookie/LoginController.java
@@ -11,7 +11,10 @@ import javafx.stage.Stage;
 
 import java.io.IOException;
 
+// Controller for the login view, handles user interaction and navigation on the login screen
 public class LoginController {
+
+    // FXML UI elements
     @FXML
     private TextField usernameField;
     @FXML
@@ -21,8 +24,10 @@ public class LoginController {
     @FXML
     private Button toHomepage;
 
+    // Access to the database
     private UserDAO userDAO = UserDAOInstance.getInstance();
 
+    // Handles login logic, authenticates the user and navigates to the homepage if successful
     @FXML
     private void handleLogin() throws IOException {
         String username = usernameField.getText();
@@ -39,6 +44,7 @@ public class LoginController {
         }
     }
 
+    // Handles navigation to registration screen when register button is clicked
     @FXML
     private void goToRegister() throws IOException {
         Stage stage = (Stage) toRegister.getScene().getWindow();
@@ -47,6 +53,7 @@ public class LoginController {
         stage.setScene(scene);
     }
 
+    // Utility method to show an alert with a message
     private void showAlert(String message) {
         Alert alert = new Alert(Alert.AlertType.INFORMATION);
         alert.setTitle("Info");
