+++ src/main/java/QUT/CAB302/fortunecookie/RegistrationController.java
@@ -11,6 +11,7 @@ import javafx.stage.Stage;
 
 import java.io.IOException;
 
+// Controller for the registration view, handles user interaction and navigation on the registration screen
 public class RegistrationController {
     @FXML
     private TextField usernameField;
@@ -19,8 +20,10 @@ public class RegistrationController {
     @FXML
     private Button toLogin;
 
+    // Access to the database
     private UserDAO userDAO = UserDAOInstance.getInstance();
 
+    // Handles registration logic, determines whether username and password is valid and navigates to login page if successful
     @FXML
     private void handleRegister() throws IOException {
         String username = usernameField.getText();
@@ -38,6 +41,7 @@ public class RegistrationController {
         }
     }
 
+    // Handles navigation to login screen when back button is clicked
     @FXML
     private void goToLogin() throws IOException {
         Stage stage = (Stage) toLogin.getScene().getWindow();
@@ -46,6 +50,7 @@ public class RegistrationController {
         stage.setScene(scene);
     }
 
+    // Utility method to show an alert with a message
     private void showAlert(String message) {
         Alert alert = new Alert(Alert.AlertType.INFORMATION);
         alert.setTitle("Info");
