+++ src/main/java/QUT/CAB302/fortunecookie/User.java
@@ -1,13 +1,19 @@
 package QUT.CAB302.fortunecookie;
 
+// A class to represent user's credentials
 public class User {
+
+    // Fields to store user variables
     private String username;
     private String password;
+
+    // Constructor to create a new user.
     public User(String username, String password){
         this.username = username;
         this.password = password;
     }
 
+    // Setters and getters for username and password
     public String getUsername() {
         return username;
     }
