+++ src/main/java/QUT/CAB302/fortunecookie/UserDAODatabase.java
@@ -1,13 +1,19 @@
 package QUT.CAB302.fortunecookie;
 
 import java.sql.*;
+
+// Implementation of the UserDAO interface using SQLite
 public class UserDAODatabase implements UserDAO {
+
+    // Database file path
     private static final String DB_URL = "jdbc:sqlite:users.db";
 
+    // Constructor for the user table
     public UserDAODatabase() {
         createUserTable();
     }
 
+    // Creates the user table if it doesn't already exist, stores username and password
     private void createUserTable() {
         try (Connection conn = DriverManager.getConnection(DB_URL);
              Statement stmt = conn.createStatement()) {
@@ -20,6 +26,8 @@ public class UserDAODatabase implements UserDAO {
             e.printStackTrace();
         }
     }
+
+    // Registers a user by inserting their credentials into the database
     @Override
     public boolean registerUser(String username, String password) {
         String sql = "INSERT INTO users(username, password) VALUES(?, ?)";
@@ -34,6 +42,8 @@ public class UserDAODatabase implements UserDAO {
             return false;
         }
     }
+
+    // Authenticates users logging in against the database
     @Override
     public User loginUser(String username, String password) {
         String sql = "SELECT * FROM users WHERE username = ? AND password = ?";
