+++ src/test/UserDAOTest.java
@@ -5,23 +5,32 @@ import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
 
+// Class containing unit tests for the login and registration logic
 public class UserDAOTest {
+
+    // Mock user data for testing
     private static final String username1 = "TestUser1";
     private static final String password1 = "TestPassword1";
     private static final String username2 = "TestUser2";
     private static final String password2 = "TestPassword2";
+
+    // DAO instance under testing
     private UserDAO userDAO;
+
+    // Initialises a fresh instance of the database before each test
     @BeforeEach
     public void setUp(){
         userDAO = new MockUserDAO(); // Tests against Mock Database as opposed to UserDAODatabase (SQL)
     }
 
+    // Tests a user can successfully be registered
     @Test
     public void testRegisterUser(){
         boolean result = userDAO.registerUser(username1, password1);
         assertTrue(result, "User successfully registered");
     }
 
+    // Tests that duplicate usernames cannot be created
     @Test
     public void testDuplicateUsers(){
         userDAO.registerUser(username1, password1);
@@ -29,6 +38,7 @@ public class UserDAOTest {
         assertFalse(result, "Duplicate Username should fail to register");
     }
 
+    // Tests that users can successfully log in
     @Test
     public void testLoginUser(){
         userDAO.registerUser(username1, password1);
@@ -37,6 +47,7 @@ public class UserDAOTest {
         assertEquals(username1, user.getUsername());
     }
 
+    // Tests that logging in with incorrect password fails
     @Test
     public void testWrongPassword(){
         userDAO.registerUser(username1, password1);
@@ -44,6 +55,7 @@ public class UserDAOTest {
         assertNull(user, "Login with wrong password should fail");
     }
 
+    // Tests that logging in with an unregistered user fails
     @Test
     public void testUserNonExistent(){
         User user = userDAO.loginUser(username1, password1);
